requirements.txt
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.13
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
httpie==2.3.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
lxml==4.2.1
MarkupSafe==1.1.1
mysqlclient==2.0.2
Pygments==2.7.4
PyJWT==1.7.1
PyMySQL==0.10.1
PySocks==1.7.1
pytz==2020.4
PyYAML==5.4.1
redis==3.5.3
requests==2.25.0
requests-toolbelt==0.9.1
selenium==3.141.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
